+++ lib/screens/sales/cash_sales_history_screen.dart
@@ -1,5 +1,6 @@
 // lib/screens/sales/cash_sales_history_screen.dart
 
+import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import '../../data/database_helper.dart';
@@ -227,7 +228,7 @@ class _CashSalesHistoryScreenState extends State<CashSalesHistoryScreen> {
   ) {
     // --- استخراج البيانات ---
     final invoiceId = invoice['InvoiceID'] as int;
-    final totalAmount = invoice['TotalAmount'] as double;
+    final totalAmount = invoice['TotalAmount'] as Decimal;
     final invoiceDate = DateTime.parse(invoice['InvoiceDate'] as String);
     final isVoid = invoice['IsVoid'] == 1;
     final status = invoice['Status'] as String?;
